#1208 Move Directive nodes from psyGen to their own files
