+++ src/psyclone/psyGen.py
@@ -47,15 +47,14 @@ import abc
 import six
 from fparser.two import Fortran2003
 from psyclone.configuration import Config
-from psyclone.f2pygen import DirectiveGen, CommentGen, CallGen, PSyIRGen, \
-    UseGen
-from psyclone.core import AccessType, VariablesAccessInfo
+from psyclone.f2pygen import CommentGen, CallGen, PSyIRGen, UseGen
+from psyclone.core import AccessType
 from psyclone.psyir.symbols import DataSymbol, ArrayType, RoutineSymbol, \
     Symbol, ContainerSymbol, GlobalInterface, INTEGER_TYPE, BOOLEAN_TYPE, \
     ArgumentInterface, DeferredType
 from psyclone.psyir.symbols.datatypes import UnknownFortranType
 from psyclone.psyir.nodes import Node, Schedule, Loop, Statement, Container, \
-    Routine, PSyDataNode, Call, StructureReference
+    Routine, Call
 from psyclone.errors import GenerationError, InternalError, FieldNotFoundError
 from psyclone.parse.algorithm import BuiltInCall
 
@@ -63,12 +62,6 @@ from psyclone.parse.algorithm import BuiltInCall
 # The types of 'intent' that an argument to a Fortran subroutine
 # may have
 FORTRAN_INTENT_NAMES = ["inout", "out", "in"]
-
-# OMP_OPERATOR_MAPPING is used to determine the operator to use in the
-# reduction clause of an OpenMP directive. All code for OpenMP
-# directives exists in psyGen.py so this mapping should not be
-# overidden.
-OMP_OPERATOR_MAPPING = {AccessType.SUM: "+"}
 
 # Mapping of access type to operator.
 REDUCTION_OPERATOR_MAPPING = {AccessType.SUM: "+"}
@@ -1162,1276 +1155,6 @@ class InvokeSchedule(Routine):
         self._opencl = value
 
 
-class Directive(Statement):
-    '''
-    Base class for all Directive statements.
-
-    All classes that generate Directive statements (e.g. OpenMP,
-    OpenACC, compiler-specific) inherit from this class.
-
-    :param ast: the entry in the fparser2 parse tree representing the code \
-                contained within this directive or None.
-    :type ast: :py:class:`fparser.two.Fortran2003.Base` or NoneType
-    :param children: list of PSyIR nodes that will be children of this \
-                     Directive node or None.
-    :type children: list of :py:class:`psyclone.psyir.nodes.Node` or NoneType
-    :param parent: PSyIR node that is the parent of this Directive or None.
-    :type parent: :py:class:`psyclone.psyir.nodes.Node` or NoneType
-
-    '''
-    # The prefix to use when constructing this directive in Fortran
-    # (e.g. "OMP"). Must be set by sub-class.
-    _PREFIX = ""
-    # Textual description of the node.
-    _children_valid_format = "Schedule"
-    _text_name = "Directive"
-    _colour = "green"
-
-    def __init__(self, ast=None, children=None, parent=None):
-        # A Directive always contains a Schedule
-        sched = self._insert_schedule(children, ast)
-        super(Directive, self).__init__(ast, children=[sched], parent=parent)
-
-    @staticmethod
-    def _validate_child(position, child):
-        '''
-        :param int position: the position to be validated.
-        :param child: a child to be validated.
-        :type child: :py:class:`psyclone.psyir.nodes.Node`
-
-        :return: whether the given child and position are valid for this node.
-        :rtype: bool
-
-        '''
-        return position == 0 and isinstance(child, Schedule)
-
-    @property
-    def dir_body(self):
-        '''
-        :returns: the Schedule associated with this directive.
-        :rtype: :py:class:`psyclone.psyir.nodes.Schedule`
-
-        :raises InternalError: if this node does not have a single Schedule as\
-                               its child.
-        '''
-        if len(self.children) != 1 or not \
-           isinstance(self.children[0], Schedule):
-            raise InternalError(
-                "Directive malformed or incomplete. It should have a single "
-                "Schedule as a child but found: {0}".format(
-                    [type(child).__name__ for child in self.children]))
-        return self.children[0]
-
-    @property
-    def dag_name(self):
-        '''
-        :returns: the name to use in the DAG for this node.
-        :rtype: str
-        '''
-        _, position = self._find_position(self.ancestor(Routine))
-        return "directive_" + str(position)
-
-    def _add_region(self, start_text, end_text=None, data_movement=None):
-        '''
-        Modifies the underlying fparser2 parse tree to include a subset
-        of nodes within a region. (e.g. a 'kernels' or 'data' region.)
-
-        :param str start_text: the directive body to insert at the \
-                               beginning of the region. "!$"+self._PREFIX+" " \
-                               is prepended to the supplied text.
-        :param str end_text: the directive body to insert at the end of \
-                             the region (or None). "!$"+self._PREFIX+" " is \
-                             prepended to the supplied text.
-        :param str data_movement: whether to include data-movement clauses and\
-                               if so, whether to determine them by analysing \
-                               the code within the region ("analyse") or to \
-                               specify 'default(present)' ("present").
-
-        :raises InternalError: if either start_text or end_text already
-                               begin with '!'.
-        :raises InternalError: if data_movement is not None and not one of \
-                               "present" or "analyse".
-        :raises InternalError: if data_movement=="analyse" and this is an \
-                               OpenMP directive.
-        '''
-        from fparser.common.readfortran import FortranStringReader
-        from fparser.two.Fortran2003 import Comment
-        from psyclone.psyir.frontend.fparser2 import Fparser2Reader
-        valid_data_movement = ["present", "analyse"]
-
-        # Ensure the fparser2 AST is up-to-date for all of our children
-        Node.update(self)
-
-        # Check that we haven't already been called
-        if self.ast:
-            return
-
-        # Sanity check the supplied begin/end text
-        if start_text.lstrip()[0] == "!":
-            raise InternalError(
-                "_add_region: start_text must be a plain label without "
-                "directive or comment characters but got: '{0}'".
-                format(start_text))
-        if end_text and end_text.lstrip()[0] == "!":
-            raise InternalError(
-                "_add_region: end_text must be a plain label without directive"
-                " or comment characters but got: '{0}'".format(end_text))
-        # We only deal with data movement if this is an OpenACC directive
-        if data_movement and data_movement == "analyse" and \
-           not isinstance(self, ACCDirective):
-            raise InternalError(
-                "_add_region: the data_movement='analyse' option is only valid"
-                " for an OpenACC directive.")
-
-        # Find a reference to the fparser2 parse tree that belongs to
-        # the contents of this region. Then go back up one level in the
-        # parse tree to find the node to which we will add directives as
-        # children. (We do this because our parent PSyIR node may be a
-        # directive which has no associated entry in the fparser2 parse tree.)
-        first_child = self.children[0][0]
-        last_child = self.children[0][-1]
-        content_ast = first_child.ast
-        fp_parent = content_ast.parent
-
-        try:
-            # Find the location of the AST of our first child node in the
-            # list of child nodes of our parent in the fparser parse tree.
-            ast_start_index = object_index(fp_parent.content,
-                                           content_ast)
-            if end_text:
-                if last_child.ast_end:
-                    ast_end_index = object_index(fp_parent.content,
-                                                 last_child.ast_end)
-                else:
-                    ast_end_index = object_index(fp_parent.content,
-                                                 last_child.ast)
-
-                text = "!$" + self._PREFIX + " " + end_text
-                directive = Comment(FortranStringReader(text,
-                                                        ignore_comments=False))
-                directive.parent = fp_parent
-                fp_parent.content.insert(ast_end_index+1, directive)
-                # Ensure this end directive is included with the set of
-                # statements belonging to this PSyIR node.
-                self.ast_end = directive
-                self.dir_body.ast_end = directive
-        except (IndexError, ValueError):
-            raise InternalError("Failed to find locations to insert "
-                                "begin/end directives.")
-
-        text = "!$" + self._PREFIX + " " + start_text
-
-        if data_movement:
-            if data_movement == "analyse":
-                # Identify the inputs and outputs to the region (variables that
-                # are read and written).
-                processor = Fparser2Reader()
-                readers, writers, readwrites = processor.get_inputs_outputs(
-                    fp_parent.content[ast_start_index:ast_end_index+1])
-
-                if readers:
-                    text += " COPYIN({0})".format(",".join(readers))
-                if writers:
-                    text += " COPYOUT({0})".format(",".join(writers))
-                if readwrites:
-                    text += " COPY({0})".format(",".join(readwrites))
-
-            elif data_movement == "present":
-                text += " DEFAULT(PRESENT)"
-            else:
-                raise InternalError(
-                    "_add_region: the optional data_movement argument must be "
-                    "one of {0} but got '{1}'".format(valid_data_movement,
-                                                      data_movement))
-        directive = Comment(FortranStringReader(text,
-                                                ignore_comments=False))
-        directive.parent = fp_parent
-        fp_parent.content.insert(ast_start_index, directive)
-
-        self.ast = directive
-        self.dir_body.ast = directive
-        # If this is a directive applied to a Loop then update the ast_end
-        # for this Node to point to the parse tree for the loop. We have to
-        # do this because the loop is a sibling (rather than a child) of the
-        # directive in the parse tree.
-        if not end_text and isinstance(first_child, Loop):
-            self.ast_end = fp_parent.content[ast_start_index+1]
-
-
-class ACCDirective(Directive):
-    ''' Base class for all OpenACC directive statements. '''
-    _PREFIX = "ACC"
-
-    @property
-    def dag_name(self):
-        ''' Return the name to use in a dag for this node.
-
-        :returns: Name of corresponding node in DAG
-        :rtype: str
-        '''
-        _, position = self._find_position(self.ancestor(Routine))
-        return "ACC_directive_" + str(position)
-
-    def validate_global_constraints(self):
-        '''
-        Perform validation checks for any global constraints. This can only
-        be done at code-generation time.
-
-        :raises GenerationError: if this ACCDirective encloses any form of \
-            PSyData node since calls to PSyData routines within OpenACC \
-            regions are not supported.
-
-        '''
-        super(ACCDirective, self).validate_global_constraints()
-
-        data_nodes = self.walk(PSyDataNode)
-        if data_nodes:
-            raise GenerationError(
-                "Cannot include calls to PSyData routines within OpenACC "
-                "regions but found {0} within a region enclosed "
-                "by an '{1}'".format(
-                    [type(node).__name__ for node in data_nodes],
-                    type(self).__name__))
-
-
-@six.add_metaclass(abc.ABCMeta)
-class ACCEnterDataDirective(ACCDirective):
-    '''
-    Abstract class representing a "!$ACC enter data" OpenACC directive in
-    an InvokeSchedule. Must be sub-classed for a particular API because the way
-    in which fields are marked as being on the remote device is API-
-    -dependent.
-
-    :param children: list of nodes which this directive should \
-                     have as children.
-    :type children: list of :py:class:`psyclone.psyir.nodes.Node`.
-    :param parent: the node in the InvokeSchedule to which to add this \
-                   directive as a child.
-    :type parent: :py:class:`psyclone.psyir.nodes.Node`.
-    '''
-    def __init__(self, children=None, parent=None):
-        super(ACCEnterDataDirective, self).__init__(children=children,
-                                                    parent=parent)
-        self._acc_dirs = None  # List of parallel directives
-
-        # The _variables_to_copy are computed dynamically until the
-        # _node_lowered flag is set to True, after that re-use the stored ones.
-        self._variables_to_copy = []
-        self._node_lowered = False
-
-    def node_str(self, colour=True):
-        '''
-        Returns the name of this node with appropriate control codes
-        to generate coloured output in a terminal that supports it.
-
-        :param bool colour: whether or not to include colour control codes.
-
-        :returns: description of this node, possibly coloured.
-        :rtype: str
-        '''
-        return self.coloured_name(colour) + "[ACC enter data]"
-
-    @property
-    def dag_name(self):
-        '''
-        :returns: the name to use for this Node in a DAG
-        :rtype: str
-        '''
-        _, position = self._find_position(self.ancestor(Routine))
-        return "ACC_data_" + str(position)
-
-    def gen_code(self, parent):
-        '''Generate the elements of the f2pygen AST for this Node in the
-        Schedule.
-
-        :param parent: node in the f2pygen AST to which to add node(s).
-        :type parent: :py:class:`psyclone.f2pygen.BaseGen`
-
-        :raises GenerationError: if no data is found to copy in.
-
-        '''
-        self.validate_global_constraints()
-
-        # We must generate a list of all of the fields accessed by
-        # OpenACC kernels (calls within an OpenACC parallel or kernels
-        # directive)
-        # 1. Find all parallel and kernels directives. We store this list for
-        #    later use in any sub-class.
-        self._acc_dirs = self.ancestor(InvokeSchedule).walk(
-                (ACCParallelDirective, ACCKernelsDirective))
-        # 2. For each directive, loop over each of the fields used by
-        #    the kernels it contains (this list is given by ref_list)
-        #    and add it to our list if we don't already have it
-        var_list = []
-        # TODO grid properties are effectively duplicated in this list (but
-        # the OpenACC deep-copy support should spot this).
-        for pdir in self._acc_dirs:
-            for var in pdir.ref_list:
-                if var not in var_list:
-                    var_list.append(var)
-        # 3. Convert this list of objects into a comma-delimited string
-        var_str = ",".join(var_list)
-        # 4. Add the enter data directive.
-        if var_str:
-            copy_in_str = "copyin("+var_str+")"
-        else:
-            # There should be at least one variable to copyin.
-            raise GenerationError(
-                "ACCEnterData directive did not find any data to copyin. "
-                "Perhaps there are no ACCParallel or ACCKernels directives "
-                "within the region.")
-        parent.add(DirectiveGen(parent, "acc", "begin", "enter data",
-                                copy_in_str))
-        # 5. Call an API-specific subclass of this class in case
-        # additional declarations are required.
-        self.data_on_device(parent)
-        parent.add(CommentGen(parent, ""))
-
-    def lower_to_language_level(self):
-        '''
-        In-place replacement of this directive concept into language level
-        PSyIR constructs.
-
-        '''
-        if not self._node_lowered:
-            # We must generate a list of all of the fields accessed by
-            # OpenACC kernels (calls within an OpenACC parallel or kernels
-            # directive)
-            # 1. Find all parallel and kernels directives. We store this list
-            # for later use in any sub-class.
-            self._acc_dirs = self.ancestor(InvokeSchedule).walk(
-                    (ACCParallelDirective, ACCKernelsDirective))
-            # 2. For each directive, loop over each of the fields used by
-            #    the kernels it contains (this list is given by ref_list)
-            #    and add it to our list if we don't already have it
-            self._variables_to_copy = []
-            # TODO grid properties are effectively duplicated in this list (but
-            # the OpenACC deep-copy support should spot this).
-            for pdir in self._acc_dirs:
-                for var in pdir.ref_list:
-                    if var not in self._variables_to_copy:
-                        self._variables_to_copy.append(var)
-            self._node_lowered = True
-
-        super(ACCEnterDataDirective, self).lower_to_language_level()
-
-    def begin_string(self):
-        '''Returns the beginning statement of this directive. The visitor is
-        responsible for adding the correct directive beginning (e.g. "!$").
-
-        :returns: the opening statement of this directive.
-        :rtype: str
-
-        '''
-        # The enter data clauses are given by the _variables_to_copy list
-        var_str = ",".join(self._variables_to_copy)
-        if var_str:
-            copy_in_str = "copyin("+var_str+")"
-        else:
-            # There should be at least one variable to copyin.
-            raise GenerationError(
-                "ACCEnterData directive did not find any data to copyin. "
-                "Perhaps there are no ACCParallel or ACCKernels directives "
-                "within the region.")
-
-        return "acc enter data " + copy_in_str
-
-    def end_string(self):
-        '''
-        :returns: the closing statement for this directive.
-        :rtype: str
-        '''
-        # pylint: disable=no-self-use
-        return ""
-
-    @abc.abstractmethod
-    def data_on_device(self, parent):
-        '''
-        Adds nodes into an InvokeSchedule to flag that the data required by the
-        kernels in the data region is now on the device.
-
-        :param parent: the node in the InvokeSchedule to which to add nodes
-        :type parent: :py:class:`psyclone.psyir.nodes.Node`
-        '''
-
-
-class ACCParallelDirective(ACCDirective):
-    '''
-    Class representing the !$ACC PARALLEL directive of OpenACC
-    in the PSyIR. By default it includes the 'DEFAULT(PRESENT)' clause which
-    means this node must either come after an EnterDataDirective or within
-    a DataDirective.
-
-    '''
-    def node_str(self, colour=True):
-        '''
-        Returns the name of this node with appropriate control codes
-        to generate coloured output in a terminal that supports it.
-
-        :param bool colour: whether or not to include colour control codes.
-
-        :returns: description of this node, possibly coloured.
-        :rtype: str
-        '''
-        return self.coloured_name(colour) + "[ACC Parallel]"
-
-    @property
-    def dag_name(self):
-        '''
-        :returns: the name to use for this Node in a DAG
-        :rtype: str
-        '''
-        _, position = self._find_position(self.ancestor(Routine))
-        return "ACC_parallel_" + str(position)
-
-    def validate_global_constraints(self):
-        '''
-        Check that the PSyIR tree containing this node is valid. Since we
-        use 'default(present)', this node must either be the child of an
-        ACCDataDirective or the parent Schedule must contain an
-        ACCEnterDataDirective before this one.
-
-        :raises GenerationError: if this ACCParallel node is not preceded by \
-            an ACCEnterDataDirective and is not the child of an \
-            ACCDataDirective.
-
-        '''
-        # We can't use Node.ancestor() because the enter data directive does
-        # not have children. Instead, we go back up to the Schedule and
-        # walk down from there.
-        routine = self.ancestor(Routine)
-        nodes = routine.walk(ACCEnterDataDirective)
-        # Check that any enter-data directive comes before this parallel
-        # directive
-        if nodes and nodes[0].abs_position > self.abs_position:
-            raise GenerationError(
-                "An ACC parallel region must be preceded by an ACC enter-"
-                "data directive but in '{0}' this is not the case.".
-                format(routine.name))
-
-        if not nodes and not self.ancestor(ACCDataDirective):
-            raise GenerationError(
-                "An ACC parallel region must either be preceded by an ACC "
-                "enter data directive or enclosed within an ACC data region "
-                "but in '{0}' this is not the case.".format(routine.name))
-
-        super(ACCParallelDirective, self).validate_global_constraints()
-
-    def gen_code(self, parent):
-        '''
-        Generate the elements of the f2pygen AST for this Node in the Schedule.
-
-        :param parent: node in the f2pygen AST to which to add node(s).
-        :type parent: :py:class:`psyclone.f2pygen.BaseGen`
-
-        '''
-        self.validate_global_constraints()
-
-        parent.add(DirectiveGen(parent, "acc", "begin", "parallel",
-                                "default(present)"))
-
-        for child in self.children:
-            child.gen_code(parent)
-
-        parent.add(DirectiveGen(parent, *self.end_string().split()))
-
-    def begin_string(self):
-        '''
-        Returns the beginning statement of this directive, i.e.
-        "acc parallel" plus any qualifiers. The backend is responsible
-        for adding the correct characters to mark this as a directive (e.g.
-        "!$").
-
-        :returns: the opening statement of this directive.
-        :rtype: str
-
-        '''
-        # pylint: disable=no-self-use
-        # "default(present)" means that the compiler is to assume that
-        # all data required by the parallel region is already present
-        # on the device. If we've made a mistake and it isn't present
-        # then we'll get a run-time error.
-        return "acc parallel default(present)"
-
-    def end_string(self):
-        '''
-        :returns: the closing statement for this directive.
-        :rtype: str
-        '''
-        # pylint: disable=no-self-use
-        return "acc end parallel"
-
-    @property
-    def ref_list(self):
-        '''
-        Returns a list of the references (whether to arrays or objects)
-        required by the Kernel call(s) that are children of this
-        directive. This is the list of quantities that must be
-        available on the remote device (probably a GPU) before
-        the parallel region can be begun.
-
-        :returns: list of variable names
-        :rtype: list of str
-        '''
-        variables = []
-
-        # Look-up the kernels that are children of this node
-        for call in self.kernels():
-            for arg in call.arguments.acc_args:
-                if arg not in variables:
-                    variables.append(arg)
-        return variables
-
-    @property
-    def fields(self):
-        '''
-        Returns a list of the names of field objects required by the Kernel
-        call(s) that are children of this directive.
-
-        :returns: list of names of field arguments.
-        :rtype: list of str
-        '''
-        # Look-up the kernels that are children of this node
-        fld_list = []
-        for call in self.kernels():
-            for arg in call.arguments.fields:
-                if arg not in fld_list:
-                    fld_list.append(arg)
-        return fld_list
-
-    @property
-    def scalars(self):
-        '''
-        Returns a list of the scalar quantities required by the Kernels in
-        this region.
-
-        :returns: list of names of scalar arguments.
-        :rtype: list of str
-        '''
-        scalars = []
-        for call in self.kernels():
-            for arg in call.arguments.scalars:
-                if arg not in scalars:
-                    scalars.append(arg)
-        return scalars
-
-    def update(self):
-        '''
-        Update the underlying fparser2 parse tree with nodes for the start
-        and end of this parallel region.
-        '''
-        self.validate_global_constraints()
-        self._add_region(start_text="PARALLEL", end_text="END PARALLEL",
-                         data_movement="present")
-
-
-class ACCLoopDirective(ACCDirective):
-    '''
-    Class managing the creation of a '!$acc loop' OpenACC directive.
-
-    :param children: list of nodes that will be children of this directive.
-    :type children: list of :py:class:`psyclone.psyir.nodes.Node`.
-    :param parent: the node in the Schedule to which to add this directive.
-    :type parent: :py:class:`psyclone.psyir.nodes.Node`.
-    :param int collapse: Number of nested loops to collapse into a single \
-                         iteration space or None.
-    :param bool independent: Whether or not to add the `independent` clause \
-                             to the loop directive.
-    '''
-    def __init__(self, children=None, parent=None, collapse=None,
-                 independent=True, sequential=False):
-        self._collapse = collapse
-        self._independent = independent
-        self._sequential = sequential
-        super(ACCLoopDirective, self).__init__(children=children,
-                                               parent=parent)
-
-    @property
-    def dag_name(self):
-        '''
-        :returns: the name to use for this Node in a DAG
-        :rtype: str
-        '''
-        _, position = self._find_position(self.ancestor(Routine))
-        return "ACC_loop_" + str(position)
-
-    def node_str(self, colour=True):
-        '''
-        Returns the name of this node with (optional) control codes
-        to generate coloured output in a terminal that supports it.
-
-        :param bool colour: whether or not to include colour control codes.
-
-        :returns: description of this node, possibly coloured.
-        :rtype: str
-        '''
-        text = self.coloured_name(colour) + "[ACC Loop"
-        if self._sequential:
-            text += ", seq"
-        else:
-            if self._collapse:
-                text += ", collapse={0}".format(self._collapse)
-            if self._independent:
-                text += ", independent"
-        text += "]"
-        return text
-
-    def validate_global_constraints(self):
-        '''
-        Perform validation of those global constraints that can only be done
-        at code-generation time.
-
-        :raises GenerationError: if this ACCLoopDirective is not enclosed \
-                            within some OpenACC parallel or kernels region.
-        '''
-        # It is only at the point of code generation that we can check for
-        # correctness (given that we don't mandate the order that a user can
-        # apply transformations to the code). As an orphaned loop directive,
-        # we must have an ACCParallelDirective or an ACCKernelsDirective as
-        # an ancestor somewhere back up the tree.
-        if not self.ancestor((ACCParallelDirective, ACCKernelsDirective)):
-            raise GenerationError(
-                "ACCLoopDirective must have an ACCParallelDirective or "
-                "ACCKernelsDirective as an ancestor in the Schedule")
-
-        super(ACCLoopDirective, self).validate_global_constraints()
-
-    def gen_code(self, parent):
-        '''
-        Generate the f2pygen AST entries in the Schedule for this OpenACC
-        loop directive.
-
-        :param parent: the parent Node in the Schedule to which to add our
-                       content.
-        :type parent: sub-class of :py:class:`psyclone.f2pygen.BaseGen`
-        :raises GenerationError: if this "!$acc loop" is not enclosed within \
-                                 an ACC Parallel region.
-        '''
-        self.validate_global_constraints()
-
-        # Add any clauses to the directive. We use self.begin_string() to avoid
-        # code duplication.
-        options_str = self.begin_string(leading_acc=False)
-
-        parent.add(DirectiveGen(parent, "acc", "begin", "loop", options_str))
-
-        for child in self.children:
-            child.gen_code(parent)
-
-    def update(self):
-        '''
-        Update the existing fparser2 parse tree with the code associated with
-        this ACC LOOP directive.
-
-        '''
-        self.validate_global_constraints()
-
-        # Use begin_string() to avoid code duplication although we have to
-        # put back the "loop" qualifier.
-        # TODO #435 remove this method altogether once the NEMO API is able to
-        # use the PSyIR backend.
-        self._add_region(
-            start_text="loop " + self.begin_string(leading_acc=False))
-
-    def begin_string(self, leading_acc=True):
-        ''' Returns the opening statement of this directive, i.e.
-        "acc loop" plus any qualifiers. If `leading_acc` is False then
-        the leading "acc loop" text is not included.
-
-        :param bool leading_acc: whether or not to include the leading \
-                                 "acc loop" in the text that is returned.
-
-        :returns: the opening statement of this directive.
-        :rtype: str
-
-        '''
-        clauses = []
-        if leading_acc:
-            clauses = ["acc", "loop"]
-
-        if self._sequential:
-            clauses.append("seq")
-        else:
-            if self._independent:
-                clauses.append("independent")
-            if self._collapse:
-                clauses.append("collapse({0})".format(self._collapse))
-        return " ".join(clauses)
-
-    def end_string(self):
-        '''
-        Would return the end string for this directive but "acc loop"
-        doesn't have a closing directive.
-
-        :returns: empty string.
-        :rtype: str
-
-        '''
-        # pylint: disable=no-self-use
-        return ""
-
-
-class OMPDirective(Directive):
-    '''
-    Base class for all OpenMP-related directives
-
-    '''
-    _PREFIX = "OMP"
-
-    @property
-    def dag_name(self):
-        '''
-        :returns: the name to use in a dag for this node
-        :rtype: str
-        '''
-        _, position = self._find_position(self.ancestor(Routine))
-        return "OMP_directive_" + str(position)
-
-    def node_str(self, colour=True):
-        '''
-        Returns the name of this node with (optional) control codes
-        to generate coloured output in a terminal that supports it.
-
-        :param bool colour: whether or not to include colour control codes.
-
-        :returns: description of this node, possibly coloured.
-        :rtype: str
-        '''
-        return self.coloured_name(colour) + "[OMP]"
-
-    def _get_reductions_list(self, reduction_type):
-        '''
-        Returns the names of all scalars within this region that require a
-        reduction of type 'reduction_type'. Returned names will be unique.
-
-        :param reduction_type: the reduction type (e.g. AccessType.SUM) to \
-                               search for.
-        :type reduction_type: :py:class:`psyclone.core.access_type.AccessType`
-
-        :returns: names of scalar arguments with reduction access.
-        :rtype: list of str
-
-        '''
-        result = []
-        const = Config.get().api_conf().get_constants()
-        for call in self.kernels():
-            for arg in call.arguments.args:
-                if arg.argument_type in const.VALID_SCALAR_NAMES:
-                    if arg.descriptor.access == reduction_type:
-                        if arg.name not in result:
-                            result.append(arg.name)
-        return result
-
-
-class OMPParallelDirective(OMPDirective):
-
-    @property
-    def dag_name(self):
-        '''
-        :returns: the name to use in the DAG for this node.
-        :rtype: str
-        '''
-        _, position = self._find_position(self.ancestor(Routine))
-        return "OMP_parallel_" + str(position)
-
-    def node_str(self, colour=True):
-        '''
-        Returns the name of this node with (optional) control codes
-        to generate coloured output in a terminal that supports it.
-
-        :param bool colour: whether or not to include colour control codes.
-
-        :returns: description of this node, possibly coloured.
-        :rtype: str
-        '''
-        return self.coloured_name(colour) + "[OMP parallel]"
-
-    def gen_code(self, parent):
-        '''Generate the fortran OMP Parallel Directive and any associated
-        code'''
-        from psyclone.f2pygen import AssignGen, UseGen, DeclGen
-
-        private_list = self._get_private_list()
-
-        reprod_red_call_list = self.reductions(reprod=True)
-        if reprod_red_call_list:
-            # we will use a private thread index variable
-            thread_idx = self.scope.symbol_table.\
-                lookup_with_tag("omp_thread_index").name
-            private_list.append(thread_idx)
-            # declare the variable
-            parent.add(DeclGen(parent, datatype="integer",
-                               entity_decls=[thread_idx]))
-        private_str = ",".join(private_list)
-
-        # We're not doing nested parallelism so make sure that this
-        # omp parallel region is not already within some parallel region
-        self.validate_global_constraints()
-
-        # Check that this OpenMP PARALLEL directive encloses other
-        # OpenMP directives. Although it is valid OpenMP if it doesn't,
-        # this almost certainly indicates a user error.
-        self._encloses_omp_directive()
-
-        calls = self.reductions()
-
-        # first check whether we have more than one reduction with the same
-        # name in this Schedule. If so, raise an error as this is not
-        # supported for a parallel region.
-        names = []
-        for call in calls:
-            name = call.reduction_arg.name
-            if name in names:
-                raise GenerationError(
-                    "Reduction variables can only be used once in an invoke. "
-                    "'{0}' is used multiple times, please use a different "
-                    "reduction variable".format(name))
-            else:
-                names.append(name)
-
-        zero_reduction_variables(calls, parent)
-
-        parent.add(DirectiveGen(parent, "omp", "begin", "parallel",
-                                "default(shared), private({0})".
-                                format(private_str)))
-
-        if reprod_red_call_list:
-            # add in a local thread index
-            parent.add(UseGen(parent, name="omp_lib", only=True,
-                              funcnames=["omp_get_thread_num"]))
-            parent.add(AssignGen(parent, lhs=thread_idx,
-                                 rhs="omp_get_thread_num()+1"))
-
-        first_type = type(self.dir_body[0])
-        for child in self.dir_body.children:
-            if first_type != type(child):
-                raise NotImplementedError("Cannot correctly generate code"
-                                          " for an OpenMP parallel region"
-                                          " containing children of "
-                                          "different types")
-            child.gen_code(parent)
-
-        parent.add(DirectiveGen(parent, "omp", "end", "parallel", ""))
-
-        if reprod_red_call_list:
-            parent.add(CommentGen(parent, ""))
-            parent.add(CommentGen(parent, " sum the partial results "
-                                  "sequentially"))
-            parent.add(CommentGen(parent, ""))
-            for call in reprod_red_call_list:
-                call.reduction_sum_loop(parent)
-
-    def begin_string(self):
-        '''Returns the beginning statement of this directive, i.e.
-        "omp parallel". The visitor is responsible for adding the
-        correct directive beginning (e.g. "!$").
-
-        :returns: the opening statement of this directive.
-        :rtype: str
-
-        '''
-        result = "omp parallel"
-        # TODO #514: not yet working with NEMO, so commented out for now
-        # if not self._reprod:
-        #     result += self._reduction_string()
-        private_list = self._get_private_list()
-        private_str = ",".join(private_list)
-
-        if private_str:
-            result = "{0} private({1})".format(result, private_str)
-        return result
-
-    def end_string(self):
-        '''Returns the end (or closing) statement of this directive, i.e.
-        "omp end parallel". The visitor is responsible for adding the
-        correct directive beginning (e.g. "!$").
-
-        :returns: the end statement for this directive.
-        :rtype: str
-
-        '''
-        # pylint: disable=no-self-use
-        return "omp end parallel"
-
-    def _get_private_list(self):
-        '''
-        Returns the variable names used for any loops within a directive
-        and any variables that have been declared private by a Kernel
-        within the directive.
-
-        :returns: list of variables to declare as thread private.
-        :rtype: list of str
-
-        :raises InternalError: if a Kernel has local variable(s) but they \
-                               aren't named.
-        '''
-        result = set()
-        # get variable names from all calls that are a child of this node
-        for call in self.kernels():
-            for variable_name in call.local_vars():
-                if variable_name == "":
-                    raise InternalError(
-                        "call '{0}' has a local variable but its "
-                        "name is not set.".format(call.name))
-                result.add(variable_name.lower())
-
-        # Now determine scalar variables that must be private:
-        var_accesses = VariablesAccessInfo()
-        self.reference_accesses(var_accesses)
-        for signature in var_accesses.all_signatures:
-            accesses = var_accesses[signature].all_accesses
-            # Ignore variables that have indices, we only look at scalar
-            if accesses[0].is_array():
-                continue
-
-            # If a variable is only accessed once, it is either an error
-            # or a shared variable - anyway it is not private
-            if len(accesses) == 1:
-                continue
-
-            # We have at least two accesses. If the first one is a write,
-            # assume the variable should be private:
-            if accesses[0].access_type == AccessType.WRITE:
-                # Check if the write access is inside the parallel loop. If
-                # the write is outside of a loop, it is an assignment to
-                # a shared variable. Example where jpk is likely used
-                # outside of the parallel section later, so it must be
-                # declared as shared in order to have its value in other loops:
-                # !$omp parallel
-                # jpk = 100
-                # !omp do
-                # do ji = 1, jpk
-
-                # TODO #598: improve the handling of scalar variables.
-
-                # Go up the tree till we either find the InvokeSchedule,
-                # which is at the top, or a Loop statement (or no parent,
-                # which means we have reached the end of a called kernel).
-                parent = accesses[0].node.ancestor((Loop, InvokeSchedule),
-                                                   include_self=True)
-
-                if parent and isinstance(parent, Loop):
-                    # The assignment to the variable is inside a loop, so
-                    # declare it to be private
-                    result.add(str(signature).lower())
-
-        # Convert the set into a list and sort it, so that we get
-        # reproducible results
-        list_result = list(result)
-        list_result.sort()
-        return list_result
-
-    def validate_global_constraints(self):
-        '''
-        Perform validation checks that can only be done at code-generation
-        time.
-
-        :raises GenerationError: if this OMPDoDirective is not enclosed \
-                            within some OpenMP parallel region.
-        '''
-        if self.ancestor(OMPParallelDirective) is not None:
-            raise GenerationError("Cannot nest OpenMP parallel regions.")
-
-    def _encloses_omp_directive(self):
-        ''' Check that this Parallel region contains other OpenMP
-            directives. While it doesn't have to (in order to be valid
-            OpenMP), it is likely that an absence of directives
-            is an error on the part of the user. '''
-        # We need to recurse down through all our children and check
-        # whether any of them are an OMPDirective.
-        node_list = self.walk(OMPDirective)
-        if not node_list:
-            # TODO raise a warning here so that the user can decide
-            # whether or not this is OK.
-            pass
-            # raise GenerationError("OpenMP parallel region does not enclose "
-            #                       "any OpenMP directives. This is probably "
-            #                       "not what you want.")
-
-    def update(self):
-        '''
-        Updates the fparser2 AST by inserting nodes for this OpenMP
-        parallel region.
-
-        '''
-        # TODO #435: Remove this function once this is fixed
-        self._add_region(
-            start_text="parallel default(shared), private({0})".format(
-                ",".join(self._get_private_list())),
-            end_text="end parallel")
-
-
-class OMPDoDirective(OMPDirective):
-    '''
-    Class representing an OpenMP DO directive in the PSyclone AST.
-
-    :param list children: list of Nodes that are children of this Node.
-    :param parent: the Node in the AST that has this directive as a child.
-    :type parent: :py:class:`psyclone.psyir.nodes.Node`
-    :param str omp_schedule: the OpenMP schedule to use.
-    :param bool reprod: whether or not to generate code for run-reproducible \
-                        OpenMP reductions.
-
-    '''
-    def __init__(self, children=None, parent=None, omp_schedule="static",
-                 reprod=None):
-
-        if children is None:
-            children = []
-
-        if reprod is None:
-            self._reprod = Config.get().reproducible_reductions
-        else:
-            self._reprod = reprod
-
-        self._omp_schedule = omp_schedule
-
-        # Call the init method of the base class once we've stored
-        # the OpenMP schedule
-        super(OMPDoDirective, self).__init__(children=children,
-                                             parent=parent)
-
-    @property
-    def dag_name(self):
-        '''
-        :returns: the name to use in the DAG for this node.
-        :rtype: str
-        '''
-        _, position = self._find_position(self.ancestor(Routine))
-        return "OMP_do_" + str(position)
-
-    def node_str(self, colour=True):
-        '''
-        Returns the name of this node with (optional) control codes
-        to generate coloured output in a terminal that supports it.
-
-        :param bool colour: whether or not to include colour control codes.
-
-        :returns: description of this node, possibly coloured.
-        :rtype: str
-        '''
-        if self.reductions():
-            reprod = "[reprod={0}]".format(self._reprod)
-        else:
-            reprod = ""
-        return "{0}[OMP do]{1}".format(self.coloured_name(colour), reprod)
-
-    def _reduction_string(self):
-        ''' Return the OMP reduction information as a string '''
-        reduction_str = ""
-        for reduction_type in AccessType.get_valid_reduction_modes():
-            reductions = self._get_reductions_list(reduction_type)
-            for reduction in reductions:
-                reduction_str += ", reduction({0}:{1})".format(
-                    OMP_OPERATOR_MAPPING[reduction_type], reduction)
-        return reduction_str
-
-    @property
-    def reprod(self):
-        ''' returns whether reprod has been set for this object or not '''
-        return self._reprod
-
-    def validate_global_constraints(self):
-        '''
-        Perform validation checks that can only be done at code-generation
-        time.
-
-        :raises GenerationError: if this OMPDoDirective is not enclosed \
-                            within some OpenMP parallel region.
-        '''
-        # It is only at the point of code generation that we can check for
-        # correctness (given that we don't mandate the order that a user
-        # can apply transformations to the code). As an orphaned loop
-        # directive, we must have an OMPParallelDirective as an ancestor
-        # somewhere back up the tree.
-        if not self.ancestor(OMPParallelDirective,
-                             excluding=OMPParallelDoDirective):
-            raise GenerationError(
-                "OMPDoDirective must be inside an OMP parallel region but "
-                "could not find an ancestor OMPParallelDirective node")
-
-        super(OMPDoDirective, self).validate_global_constraints()
-
-    def gen_code(self, parent):
-        '''
-        Generate the f2pygen AST entries in the Schedule for this OpenMP do
-        directive.
-
-        :param parent: the parent Node in the Schedule to which to add our \
-                       content.
-        :type parent: sub-class of :py:class:`psyclone.f2pygen.BaseGen`
-        :raises GenerationError: if this "!$omp do" is not enclosed within \
-                                 an OMP Parallel region.
-
-        '''
-        self.validate_global_constraints()
-
-        if self._reprod:
-            local_reduction_string = ""
-        else:
-            local_reduction_string = self._reduction_string()
-
-        # As we're an orphaned loop we don't specify the scope
-        # of any variables so we don't have to generate the
-        # list of private variables
-        options = "schedule({0})".format(self._omp_schedule) + \
-                  local_reduction_string
-        parent.add(DirectiveGen(parent, "omp", "begin", "do", options))
-
-        for child in self.children:
-            child.gen_code(parent)
-
-        # make sure the directive occurs straight after the loop body
-        position = parent.previous_loop()
-        parent.add(DirectiveGen(parent, "omp", "end", "do", ""),
-                   position=["after", position])
-
-    def begin_string(self):
-        '''Returns the beginning statement of this directive, i.e.
-        "omp do ...". The visitor is responsible for adding the
-        correct directive beginning (e.g. "!$").
-
-        :returns: the beginning statement for this directive.
-        :rtype: str
-
-        '''
-        return "omp do schedule({0})".format(self._omp_schedule)
-
-    def end_string(self):
-        '''Returns the end (or closing) statement of this directive, i.e.
-        "omp end do". The visitor is responsible for adding the
-        correct directive beginning (e.g. "!$").
-
-        :returns: the end statement for this directive.
-        :rtype: str
-
-        '''
-        # pylint: disable=no-self-use
-        return "omp end do"
-
-    def update(self):
-        '''
-        Updates the fparser2 AST by inserting nodes for this OpenMP do.
-
-        :raises GenerationError: if the existing AST doesn't have the \
-                                 correct structure to permit the insertion \
-                                 of the OpenMP parallel do.
-        '''
-        self.validate_global_constraints()
-
-        # Since this is an OpenMP do, it can only be applied
-        # to a single loop.
-        if len(self.dir_body.children) != 1:
-            raise GenerationError(
-                "An OpenMP DO can only be applied to a single loop "
-                "but this Node has {0} children: {1}".
-                format(len(self.dir_body.children), self.dir_body.children))
-
-        self._add_region(start_text="do schedule({0})".format(
-            self._omp_schedule), end_text="end do")
-
-
-class OMPParallelDoDirective(OMPParallelDirective, OMPDoDirective):
-    ''' Class for the !$OMP PARALLEL DO directive. This inherits from
-        both OMPParallelDirective (because it creates a new OpenMP
-        thread-parallel region) and OMPDoDirective (because it
-        causes a loop to be parallelised). '''
-
-    def __init__(self, children=[], parent=None, omp_schedule="static"):
-        OMPDoDirective.__init__(self,
-                                children=children,
-                                parent=parent,
-                                omp_schedule=omp_schedule)
-
-    @property
-    def dag_name(self):
-        '''
-        :returns: the name to use in the DAG for this node.
-        :rtype: str
-        '''
-        _, position = self._find_position(self.ancestor(Routine))
-        return "OMP_parallel_do_" + str(position)
-
-    def node_str(self, colour=True):
-        '''
-        Returns the name of this node with (optional) control codes
-        to generate coloured output in a terminal that supports it.
-
-        :param bool colour: whether or not to include colour control codes.
-
-        :returns: description of this node, possibly coloured.
-        :rtype: str
-        '''
-        return self.coloured_name(colour) + "[OMP parallel do]"
-
-    def gen_code(self, parent):
-
-        # We're not doing nested parallelism so make sure that this
-        # omp parallel do is not already within some parallel region
-        self.validate_global_constraints()
-
-        calls = self.reductions()
-        zero_reduction_variables(calls, parent)
-        private_str = ",".join(self._get_private_list())
-        parent.add(DirectiveGen(parent, "omp", "begin", "parallel do",
-                                "default(shared), private({0}), "
-                                "schedule({1})".
-                                format(private_str, self._omp_schedule) +
-                                self._reduction_string()))
-        for child in self.children:
-            child.gen_code(parent)
-
-        # make sure the directive occurs straight after the loop body
-        position = parent.previous_loop()
-        parent.add(DirectiveGen(parent, *self.end_string().split()),
-                   position=["after", position])
-
-    def begin_string(self):
-        '''Returns the beginning statement of this directive, i.e.
-        "omp do ...". The visitor is responsible for adding the
-        correct directive beginning (e.g. "!$").
-
-        :returns: the beginning statement for this directive.
-        :rtype: str
-
-        '''
-        private_str = ",".join(self._get_private_list())
-        return ("omp parallel do default(shared), private({0}), "
-                "schedule({1})".format(private_str, self._omp_schedule) +
-                self._reduction_string())
-
-    def end_string(self):
-        '''
-        :returns: the closing statement for this directive.
-        :rtype: str
-        '''
-        # pylint: disable=no-self-use
-        return "omp end parallel do"
-
-    def update(self):
-        '''
-        Updates the fparser2 AST by inserting nodes for this OpenMP
-        parallel do.
-
-        :raises GenerationError: if the existing AST doesn't have the \
-                                 correct structure to permit the insertion \
-                                 of the OpenMP parallel do.
-        '''
-        # Since this is an OpenMP (parallel) do, it can only be applied
-        # to a single loop.
-        if len(self.dir_body.children) != 1:
-            raise GenerationError(
-                "An OpenMP PARALLEL DO can only be applied to a single loop "
-                "but this Node has {0} children: {1}".
-                format(len(self.dir_body.children), self.dir_body.children))
-
-        self._add_region(
-            start_text="parallel do default(shared), private({0}), "
-            "schedule({1})".format(",".join(self._get_private_list()),
-                                   self._omp_schedule),
-            end_text="end parallel do")
-
-
 class GlobalSum(Statement):
     '''
     Generic Global Sum class which can be added to and manipulated
@@ -2767,6 +1490,7 @@ class Kern(Statement):
         do loop. If so report whether it has the reproducible flag
         set. Note, this also catches OMPParallelDo Directives but they
         have reprod set to False so it is OK.'''
+        from psyclone.psyir.nodes import OMPDoDirective
         ancestor = self.ancestor(OMPDoDirective)
         if ancestor:
             return ancestor.reprod
@@ -4592,265 +3316,8 @@ class DummyTransformation(Transformation):
         return None, None
 
 
-class ACCKernelsDirective(ACCDirective):
-    '''
-    Class representing the !$ACC KERNELS directive in the PSyIR.
-
-    :param children: the PSyIR nodes to be enclosed in the Kernels region \
-                     and which are therefore children of this node.
-    :type children: list of sub-classes of \
-                    :py:class:`psyclone.psyir.nodes.Node`
-    :param parent: the parent of this node in the PSyIR.
-    :type parent: sub-class of :py:class:`psyclone.psyir.nodes.Node`
-    :param bool default_present: whether or not to add the "default(present)" \
-                                 clause to the kernels directive.
-
-    :raises NotImplementedError: if default_present is False.
-
-    '''
-    def __init__(self, children=None, parent=None, default_present=True):
-        super(ACCKernelsDirective, self).__init__(children=children,
-                                                  parent=parent)
-        self._default_present = default_present
-
-    @property
-    def dag_name(self):
-        '''
-        :returns: the name to use for this node in a dag.
-        :rtype: str
-        '''
-        _, position = self._find_position(self.ancestor(Routine))
-        return "ACC_kernels_" + str(position)
-
-    def node_str(self, colour=True):
-        ''' Returns the name of this node with (optional) control codes
-        to generate coloured output in a terminal that supports it.
-
-        :param bool colour: whether or not to include colour control codes.
-
-        :returns: description of this node, possibly coloured.
-        :rtype: str
-        '''
-        return self.coloured_name(colour) + "[ACC Kernels]"
-
-    def gen_code(self, parent):
-        '''
-        Generate the f2pygen AST entries in the Schedule for this
-        OpenACC Kernels directive.
-
-        :param parent: the parent Node in the Schedule to which to add this \
-                       content.
-        :type parent: sub-class of :py:class:`psyclone.f2pygen.BaseGen`
-
-        '''
-        self.validate_global_constraints()
-
-        # We re-use the 'begin_string' method but must skip the leading 'acc'
-        # that it includes.
-        parent.add(DirectiveGen(parent, "acc", "begin",
-                                *self.begin_string().split()[1:]))
-        for child in self.children:
-            child.gen_code(parent)
-        parent.add(DirectiveGen(parent, *self.end_string().split()))
-
-    @property
-    def ref_list(self):
-        '''
-        Returns a list of the references (whether to arrays or objects)
-        required by the Kernel call(s) that are children of this
-        directive. This is the list of quantities that must be
-        available on the remote device (probably a GPU) before
-        the parallel region can be begun.
-
-        :returns: list of variable names
-        :rtype: list of str
-        '''
-        variables = []
-
-        # Look-up the kernels that are children of this node
-        for call in self.kernels():
-            for arg in call.arguments.acc_args:
-                if arg not in variables:
-                    variables.append(arg)
-        return variables
-
-    def begin_string(self):
-        '''Returns the beginning statement of this directive, i.e.
-        "acc kernels ...". The backend is responsible for adding the
-        correct directive beginning (e.g. "!$").
-
-        :returns: the beginning statement for this directive.
-        :rtype: str
-
-        '''
-        result = "acc kernels"
-        if self._default_present:
-            result += " default(present)"
-        return result
-
-    def end_string(self):
-        '''
-        Returns the ending statement for this directive. The backend is
-        responsible for adding the language-specific syntax that marks this
-        as a directive.
-
-        :returns: the closing statement for this directive.
-        :rtype: str
-
-        '''
-        # pylint: disable=no-self-use
-        return "acc end kernels"
-
-    def update(self):
-        '''
-        Updates the fparser2 AST by inserting nodes for this ACC kernels
-        directive.
-
-        TODO #435 remove this routine once the NEMO API is able to use the
-        PSyIR backend.
-
-        '''
-        self.validate_global_constraints()
-
-        data_movement = None
-        if self._default_present:
-            data_movement = "present"
-        self._add_region(start_text="KERNELS", end_text="END KERNELS",
-                         data_movement=data_movement)
-
-
-class ACCDataDirective(ACCDirective):
-    '''
-    Class representing the !$ACC DATA ... !$ACC END DATA directive
-    in the PSyIR.
-
-    '''
-    @property
-    def dag_name(self):
-        '''
-        :returns: the name to use in a dag for this node.
-        :rtype: str
-        '''
-        _, position = self._find_position(self.ancestor(Routine))
-        return "ACC_data_" + str(position)
-
-    def node_str(self, colour=True):
-        ''' Returns the name of this node with (optional) control codes
-        to generate coloured output in a terminal that supports it.
-
-        :param bool colour: whether or not to include colour control codes.
-
-        :returns: description of this node, possibly coloured.
-        :rtype: str
-        '''
-        return self.coloured_name(colour) + "[ACC DATA]"
-
-    def gen_code(self, _):
-        '''
-        :raises InternalError: the ACC data directive is currently only \
-                               supported for the NEMO API and that uses the \
-                               update() method to alter the underlying \
-                               fparser2 parse tree.
-
-        TODO #435 update above explanation when update() method is removed.
-
-        '''
-        raise InternalError(
-            "ACCDataDirective.gen_code should not have been called.")
-
-    def update(self):
-        '''
-        Updates the fparser2 AST by inserting nodes for this OpenACC Data
-        directive.
-
-        '''
-        self.validate_global_constraints()
-        self._add_region(start_text="DATA", end_text="END DATA",
-                         data_movement="analyse")
-
-    def begin_string(self):
-        '''Returns the beginning statement of this directive, i.e.
-        "acc data". The backend is responsible for adding the
-        correct directive beginning (e.g. "!$").
-
-        :returns: the opening statement of this directive.
-        :rtype: str
-
-        :raises NotImplementedError: if the region contains one or more \
-                                     references to structures (TODO #1028).
-
-        '''
-        result = "acc data"
-
-        struct_accesses = self.walk(StructureReference)
-        if struct_accesses:
-            # TODO #1028. Dependence analysis does not yet work for structure
-            # references.
-            # Have to import here to avoid circular dependency
-            # pylint: disable=import-outside-toplevel
-            from psyclone.psyir.backend.fortran import FortranWriter
-            fwriter = FortranWriter()
-            ref_list = [fwriter(ref) for ref in struct_accesses]
-            raise NotImplementedError(
-                "Structure (derived-type) references are not yet supported "
-                "within OpenACC data regions but found: {0}".format(ref_list))
-
-        # Identify the inputs and outputs to the region (variables that
-        # are read and written).
-        var_accesses = VariablesAccessInfo(self)
-        table = self.scope.symbol_table
-        readers = set()
-        writers = set()
-        for signature in var_accesses.all_signatures:
-            var = str(signature)
-            sym = table.lookup(var)
-            accesses = var_accesses[signature]
-            if not sym.is_array:
-                # We ignore scalars
-                continue
-            if accesses.is_read():
-                readers.add(signature)
-            if accesses.is_written():
-                writers.add(signature)
-        readwrites = readers.intersection(writers)
-        # Are any of the read-writes written before they are read?
-        for signature in list(readwrites)[:]:
-            accesses = var_accesses[signature]
-            if accesses[0].access_type == AccessType.WRITE:
-                # First access is a write so treat as a write
-                writers.add(signature)
-                readers.discard(signature)
-                readwrites.discard(signature)
-        readers_list = sorted(list(readers - readwrites))
-        writers_list = sorted(list(writers - readwrites))
-        readwrites_list = sorted(list(readwrites))
-        if readers_list:
-            str_readers = [str(sig) for sig in readers_list]
-            result += " copyin({0})".format(",".join(str_readers))
-        if writers_list:
-            str_writers = [str(sig) for sig in writers_list]
-            result += " copyout({0})".format(",".join(str_writers))
-        if readwrites_list:
-            str_readwrites = [str(sig) for sig in readwrites_list]
-            result += " copy({0})".format(",".join(str_readwrites))
-
-        return result
-
-    def end_string(self):
-        '''
-        :returns: the text for the end of this directive region.
-        :rtype: str
-
-        '''
-        # pylint: disable=no-self-use
-        return "acc end data"
-
-
 # For Sphinx AutoAPI documentation generation
 __all__ = ['PSyFactory', 'PSy', 'Invokes', 'Invoke', 'InvokeSchedule',
-           'Directive', 'ACCDirective', 'ACCEnterDataDirective',
-           'ACCParallelDirective', 'ACCLoopDirective', 'OMPDirective',
-           'OMPParallelDirective', 'OMPDoDirective', 'OMPParallelDoDirective',
            'GlobalSum', 'HaloExchange', 'Kern', 'CodedKern', 'InlinedKern',
            'BuiltIn', 'Arguments', 'DataAccess', 'Argument', 'KernelArgument',
            'TransInfo', 'Transformation', 'DummyTransformation',
