+++ src/psyclone/tests/nemo/transformations/openacc/kernels_directive_test.py
@@ -43,7 +43,8 @@ import pytest
 
 from fparser.common.readfortran import FortranStringReader
 
-from psyclone.psyGen import ACCKernelsDirective, Loop, PSyFactory
+from psyclone.psyGen import PSyFactory
+from psyclone.psyir.nodes import  ACCKernelsDirective, Loop
 from psyclone.psyir.transformations import TransformationError, ProfileTrans
 from psyclone.transformations import ACCKernelsTrans, ACCLoopTrans
 from psyclone.psyir.nodes import Assignment
