+++ src/psyclone/tests/nemo/transformations/openacc/parallel_directive_test.py
@@ -40,7 +40,8 @@
 
 from __future__ import print_function, absolute_import
 from fparser.common.readfortran import FortranStringReader
-from psyclone.psyGen import PSyFactory, TransInfo, ACCParallelDirective
+from psyclone.psyGen import PSyFactory, TransInfo
+from psyclone.psyir.nodes import ACCParallelDirective
 
 
 # The PSyclone API under test
